refactor: move project store into domain package
