+++ pkg/agentcompose/project_store.go
@@ -2,930 +2,150 @@ package agentcompose
 
 import (
 	"context"
-	"database/sql"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"strings"
-	"time"
 
+	projectdomain "agent-compose/internal/agentcompose/project"
 	"agent-compose/pkg/agentcompose/domain"
 	"agent-compose/pkg/compose"
 )
 
 const (
-	ProjectRunStatusPending   = domain.ProjectRunStatusPending
-	ProjectRunStatusRunning   = domain.ProjectRunStatusRunning
-	ProjectRunStatusSucceeded = domain.ProjectRunStatusSucceeded
-	ProjectRunStatusFailed    = domain.ProjectRunStatusFailed
-	ProjectRunStatusCanceled  = domain.ProjectRunStatusCanceled
+	ProjectRunStatusPending   = projectdomain.ProjectRunStatusPending
+	ProjectRunStatusRunning   = projectdomain.ProjectRunStatusRunning
+	ProjectRunStatusSucceeded = projectdomain.ProjectRunStatusSucceeded
+	ProjectRunStatusFailed    = projectdomain.ProjectRunStatusFailed
+	ProjectRunStatusCanceled  = projectdomain.ProjectRunStatusCanceled
 
-	ProjectRunSourceManual    = domain.ProjectRunSourceManual
-	ProjectRunSourceScheduler = domain.ProjectRunSourceScheduler
-	ProjectRunSourceAPI       = domain.ProjectRunSourceAPI
+	ProjectRunSourceManual    = projectdomain.ProjectRunSourceManual
+	ProjectRunSourceScheduler = projectdomain.ProjectRunSourceScheduler
+	ProjectRunSourceAPI       = projectdomain.ProjectRunSourceAPI
 )
 
-type ProjectRecord struct {
-	ID              string    `json:"id"`
-	Name            string    `json:"name"`
-	SourcePath      string    `json:"source_path,omitempty"`
-	SourceJSON      string    `json:"source_json"`
-	CurrentRevision int64     `json:"current_revision"`
-	SpecHash        string    `json:"spec_hash,omitempty"`
-	CreatedAt       time.Time `json:"created_at"`
-	UpdatedAt       time.Time `json:"updated_at"`
-	RemovedAt       time.Time `json:"removed_at,omitempty"`
-}
-
-type ProjectRevisionRecord struct {
-	ProjectID string    `json:"project_id"`
-	Revision  int64     `json:"revision"`
-	SpecHash  string    `json:"spec_hash"`
-	SpecJSON  string    `json:"spec_json"`
-	CreatedAt time.Time `json:"created_at"`
-}
-
-type ProjectAgentRecord struct {
-	ProjectID        string    `json:"project_id"`
-	AgentName        string    `json:"agent_name"`
-	ManagedAgentID   string    `json:"managed_agent_id,omitempty"`
-	Revision         int64     `json:"revision"`
-	Provider         string    `json:"provider,omitempty"`
-	Model            string    `json:"model,omitempty"`
-	Image            string    `json:"image,omitempty"`
-	Driver           string    `json:"driver,omitempty"`
-	SchedulerEnabled bool      `json:"scheduler_enabled"`
-	SpecJSON         string    `json:"spec_json"`
-	CreatedAt        time.Time `json:"created_at"`
-	UpdatedAt        time.Time `json:"updated_at"`
-}
-
-type ProjectSchedulerRecord struct {
-	ProjectID       string    `json:"project_id"`
-	SchedulerID     string    `json:"scheduler_id"`
-	AgentName       string    `json:"agent_name"`
-	ManagedLoaderID string    `json:"managed_loader_id,omitempty"`
-	Revision        int64     `json:"revision"`
-	Enabled         bool      `json:"enabled"`
-	TriggerCount    int       `json:"trigger_count"`
-	SpecJSON        string    `json:"spec_json"`
-	CreatedAt       time.Time `json:"created_at"`
-	UpdatedAt       time.Time `json:"updated_at"`
-}
-
-type ProjectRunRecord struct {
-	RunID           string    `json:"run_id"`
-	ProjectID       string    `json:"project_id"`
-	ProjectName     string    `json:"project_name,omitempty"`
-	ProjectRevision int64     `json:"project_revision"`
-	AgentName       string    `json:"agent_name,omitempty"`
-	ManagedAgentID  string    `json:"managed_agent_id,omitempty"`
-	Source          string    `json:"source,omitempty"`
-	SchedulerID     string    `json:"scheduler_id,omitempty"`
-	TriggerID       string    `json:"trigger_id,omitempty"`
-	Status          string    `json:"status"`
-	SessionID       string    `json:"session_id,omitempty"`
-	ExitCode        int       `json:"exit_code,omitempty"`
-	Error           string    `json:"error,omitempty"`
-	Prompt          string    `json:"prompt,omitempty"`
-	Output          string    `json:"output,omitempty"`
-	ResultJSON      string    `json:"result_json,omitempty"`
-	LogsPath        string    `json:"logs_path,omitempty"`
-	ArtifactsDir    string    `json:"artifacts_dir,omitempty"`
-	CleanupError    string    `json:"cleanup_error,omitempty"`
-	Driver          string    `json:"driver,omitempty"`
-	ImageRef        string    `json:"image_ref,omitempty"`
-	StartedAt       time.Time `json:"started_at,omitempty"`
-	CompletedAt     time.Time `json:"completed_at,omitempty"`
-	DurationMs      int64     `json:"duration_ms,omitempty"`
-	CreatedAt       time.Time `json:"created_at"`
-	UpdatedAt       time.Time `json:"updated_at"`
-}
-
-type ProjectListOptions struct {
-	Query          string
-	IncludeRemoved bool
-	Offset         int
-	Limit          int
-}
-
-type ProjectRunListOptions struct {
-	ProjectID   string
-	AgentName   string
-	SessionID   string
-	SchedulerID string
-	Status      string
-	Source      string
-	Offset      int
-	Limit       int
-}
-
-type ProjectListResult struct {
-	Projects   []ProjectRecord
-	TotalCount int
-	HasMore    bool
-	NextOffset int
-}
+type ProjectRecord = projectdomain.ProjectRecord
+type ProjectRevisionRecord = projectdomain.ProjectRevisionRecord
+type ProjectAgentRecord = projectdomain.ProjectAgentRecord
+type ProjectSchedulerRecord = projectdomain.ProjectSchedulerRecord
+type ProjectRunRecord = projectdomain.ProjectRunRecord
+type ProjectListOptions = projectdomain.ProjectListOptions
+type ProjectRunListOptions = projectdomain.ProjectRunListOptions
+type ProjectListResult = projectdomain.ProjectListResult
 
 func StableProjectID(name, sourcePath string) (string, error) {
-	return domain.StableProjectID(name, sourcePath)
+	return projectdomain.StableProjectID(name, sourcePath)
 }
 
 func StableManagedAgentID(projectID, agentName string) (string, error) {
-	return domain.StableManagedAgentID(projectID, agentName)
+	return projectdomain.StableManagedAgentID(projectID, agentName)
 }
 
 func StableProjectSchedulerID(projectID, agentName, schedulerName string) (string, error) {
-	return domain.StableProjectSchedulerID(projectID, agentName, schedulerName)
+	return projectdomain.StableProjectSchedulerID(projectID, agentName, schedulerName)
 }
 
 func StableManagedLoaderID(projectID, agentName, schedulerName string) (string, error) {
-	return domain.StableManagedLoaderID(projectID, agentName, schedulerName)
+	return projectdomain.StableManagedLoaderID(projectID, agentName, schedulerName)
 }
 
 func StableManagedTriggerID(projectID, agentName, schedulerName, triggerName string, triggerIndex int) (string, error) {
-	return domain.StableManagedTriggerID(projectID, agentName, schedulerName, triggerName, triggerIndex)
+	return projectdomain.StableManagedTriggerID(projectID, agentName, schedulerName, triggerName, triggerIndex)
 }
 
 func StableProjectRunID(projectID, agentName, source, idempotencyKey string) (string, error) {
-	return domain.StableProjectRunID(projectID, agentName, source, idempotencyKey)
+	return projectdomain.StableProjectRunID(projectID, agentName, source, idempotencyKey)
 }
 
 func NewProjectRecordFromSpec(spec *compose.NormalizedProjectSpec, sourcePath string) (ProjectRecord, error) {
-	if spec == nil {
-		return ProjectRecord{}, fmt.Errorf("project spec is required")
-	}
-	sourcePath = normalizeProjectSourcePath(sourcePath)
-	projectID, err := StableProjectID(spec.Name, sourcePath)
-	if err != nil {
-		return ProjectRecord{}, err
-	}
-	specHash, err := spec.Hash()
-	if err != nil {
-		return ProjectRecord{}, fmt.Errorf("hash project spec: %w", err)
-	}
-	sourceJSON, err := encodeProjectSourceJSON(sourcePath)
-	if err != nil {
-		return ProjectRecord{}, err
-	}
-	return ProjectRecord{
-		ID:         projectID,
-		Name:       strings.TrimSpace(spec.Name),
-		SourcePath: sourcePath,
-		SourceJSON: sourceJSON,
-		SpecHash:   specHash,
-	}, nil
+	return projectdomain.NewProjectRecordFromSpec(spec, sourcePath)
 }
 
 func NewProjectAgentRecordFromSpec(projectID string, revision int64, agent compose.NormalizedAgentSpec) (ProjectAgentRecord, error) {
-	managedAgentID, err := StableManagedAgentID(projectID, agent.Name)
-	if err != nil {
-		return ProjectAgentRecord{}, err
-	}
-	specJSON, err := marshalCanonicalProjectJSON(agent)
-	if err != nil {
-		return ProjectAgentRecord{}, fmt.Errorf("marshal project agent %s spec: %w", agent.Name, err)
-	}
-	driver := ""
-	if agent.Driver != nil {
-		driver = agent.Driver.Name
-	}
-	return ProjectAgentRecord{
-		ProjectID:        strings.TrimSpace(projectID),
-		AgentName:        strings.TrimSpace(agent.Name),
-		ManagedAgentID:   managedAgentID,
-		Revision:         revision,
-		Provider:         strings.TrimSpace(agent.Provider),
-		Model:            strings.TrimSpace(agent.Model),
-		Image:            strings.TrimSpace(agent.Image),
-		Driver:           strings.TrimSpace(driver),
-		SchedulerEnabled: agent.Scheduler != nil && agent.Scheduler.Enabled,
-		SpecJSON:         string(specJSON),
-	}, nil
+	return projectdomain.NewProjectAgentRecordFromSpec(projectID, revision, agent)
 }
 
 func NewProjectSchedulerRecordFromSpec(projectID string, revision int64, agent compose.NormalizedAgentSpec) (ProjectSchedulerRecord, bool, error) {
-	if agent.Scheduler == nil {
-		return ProjectSchedulerRecord{}, false, nil
-	}
-	schedulerID, err := StableProjectSchedulerID(projectID, agent.Name, "")
-	if err != nil {
-		return ProjectSchedulerRecord{}, false, err
-	}
-	loaderID, err := StableManagedLoaderID(projectID, agent.Name, "")
-	if err != nil {
-		return ProjectSchedulerRecord{}, false, err
-	}
-	specJSON, err := marshalCanonicalProjectJSON(agent.Scheduler)
-	if err != nil {
-		return ProjectSchedulerRecord{}, false, fmt.Errorf("marshal project scheduler %s spec: %w", agent.Name, err)
-	}
-	return ProjectSchedulerRecord{
-		ProjectID:       strings.TrimSpace(projectID),
-		SchedulerID:     schedulerID,
-		AgentName:       strings.TrimSpace(agent.Name),
-		ManagedLoaderID: loaderID,
-		Revision:        revision,
-		Enabled:         agent.Scheduler.Enabled,
-		TriggerCount:    len(agent.Scheduler.Triggers),
-		SpecJSON:        string(specJSON),
-	}, true, nil
+	return projectdomain.NewProjectSchedulerRecordFromSpec(projectID, revision, agent)
+}
+
+func (s *ConfigStore) projectStore() *projectdomain.Store {
+	return projectdomain.NewStore(s.db)
 }
 
 func (s *ConfigStore) UpsertProject(ctx context.Context, project ProjectRecord) (ProjectRecord, error) {
-	project, err := normalizeProjectRecord(project)
-	if err != nil {
-		return ProjectRecord{}, err
-	}
-	now := time.Now().UTC()
-	existing, found, err := s.getProject(ctx, project.ID, true)
-	if err != nil {
-		return ProjectRecord{}, err
-	}
-	if found {
-		project.CreatedAt = existing.CreatedAt
-		project.CurrentRevision = existing.CurrentRevision
-		if project.SpecHash == "" {
-			project.SpecHash = existing.SpecHash
-		}
-		project.UpdatedAt = now
-		project.RemovedAt = time.Time{}
-		result, err := s.db.ExecContext(ctx, `UPDATE project SET
-			name = ?, source_path = ?, source_json = ?, spec_hash = ?, updated_at = ?, removed_at = 0
-			WHERE id = ?`,
-			project.Name, project.SourcePath, project.SourceJSON, project.SpecHash, project.UpdatedAt.Unix(), project.ID)
-		if err != nil {
-			return ProjectRecord{}, fmt.Errorf("update project %s: %w", project.ID, err)
-		}
-		if rows, _ := result.RowsAffected(); rows == 0 {
-			return ProjectRecord{}, fmt.Errorf("project %s not found", project.ID)
-		}
-		return s.GetProject(ctx, project.ID)
-	}
-
-	project.CreatedAt = now
-	project.UpdatedAt = now
-	if _, err := s.db.ExecContext(ctx, `INSERT INTO project(
-		id, name, source_path, source_json, current_revision, spec_hash, created_at, updated_at, removed_at
-	) VALUES(?, ?, ?, ?, ?, ?, ?, ?, 0)`,
-		project.ID, project.Name, project.SourcePath, project.SourceJSON, project.CurrentRevision, project.SpecHash, project.CreatedAt.Unix(), project.UpdatedAt.Unix()); err != nil {
-		return ProjectRecord{}, fmt.Errorf("insert project %s: %w", project.ID, err)
-	}
-	return s.GetProject(ctx, project.ID)
+	return s.projectStore().UpsertProject(ctx, project)
 }
 
 func (s *ConfigStore) SaveProjectRevision(ctx context.Context, revision ProjectRevisionRecord) (ProjectRevisionRecord, bool, error) {
-	revision.ProjectID = strings.TrimSpace(revision.ProjectID)
-	revision.SpecHash = strings.TrimSpace(revision.SpecHash)
-	revision.SpecJSON = strings.TrimSpace(revision.SpecJSON)
-	if revision.ProjectID == "" || revision.SpecHash == "" || revision.SpecJSON == "" {
-		return ProjectRevisionRecord{}, false, fmt.Errorf("project id, spec hash, and spec json are required")
-	}
-	if !json.Valid([]byte(revision.SpecJSON)) {
-		return ProjectRevisionRecord{}, false, fmt.Errorf("project revision spec_json must be valid JSON")
-	}
-
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return ProjectRevisionRecord{}, false, fmt.Errorf("begin project revision tx: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	row := tx.QueryRowContext(ctx, `SELECT project_id, revision, spec_hash, spec_json, created_at
-		FROM project_revision WHERE project_id = ? AND spec_hash = ?`, revision.ProjectID, revision.SpecHash)
-	existing, err := scanProjectRevision(row.Scan)
-	if err == nil {
-		return existing, false, tx.Commit()
-	}
-	if !errors.Is(err, sql.ErrNoRows) {
-		return ProjectRevisionRecord{}, false, err
-	}
-
-	var nextRevision int64
-	if err := tx.QueryRowContext(ctx, `SELECT COALESCE(MAX(revision), 0) + 1 FROM project_revision WHERE project_id = ?`, revision.ProjectID).Scan(&nextRevision); err != nil {
-		return ProjectRevisionRecord{}, false, fmt.Errorf("query next project revision %s: %w", revision.ProjectID, err)
-	}
-	now := time.Now().UTC()
-	revision.Revision = nextRevision
-	revision.CreatedAt = now
-	if _, err := tx.ExecContext(ctx, `INSERT INTO project_revision(project_id, revision, spec_hash, spec_json, created_at)
-		VALUES(?, ?, ?, ?, ?)`, revision.ProjectID, revision.Revision, revision.SpecHash, revision.SpecJSON, revision.CreatedAt.Unix()); err != nil {
-		return ProjectRevisionRecord{}, false, fmt.Errorf("insert project revision %s/%d: %w", revision.ProjectID, revision.Revision, err)
-	}
-	result, err := tx.ExecContext(ctx, `UPDATE project SET current_revision = ?, spec_hash = ?, updated_at = ?, removed_at = 0 WHERE id = ?`,
-		revision.Revision, revision.SpecHash, now.Unix(), revision.ProjectID)
-	if err != nil {
-		return ProjectRevisionRecord{}, false, fmt.Errorf("update project revision pointer %s: %w", revision.ProjectID, err)
-	}
-	if rows, _ := result.RowsAffected(); rows == 0 {
-		return ProjectRevisionRecord{}, false, fmt.Errorf("project %s not found", revision.ProjectID)
-	}
-	if err := tx.Commit(); err != nil {
-		return ProjectRevisionRecord{}, false, fmt.Errorf("commit project revision tx: %w", err)
-	}
-	return revision, true, nil
+	return s.projectStore().SaveProjectRevision(ctx, revision)
 }
 
 func (s *ConfigStore) GetProject(ctx context.Context, projectID string) (ProjectRecord, error) {
-	item, found, err := s.getProject(ctx, projectID, false)
-	if err != nil {
-		return ProjectRecord{}, err
-	}
-	if !found {
-		return ProjectRecord{}, fmt.Errorf("project %s not found: %w", strings.TrimSpace(projectID), sql.ErrNoRows)
-	}
-	return item, nil
+	return s.projectStore().GetProject(ctx, projectID)
 }
 
 func (s *ConfigStore) ListProjects(ctx context.Context, options ProjectListOptions) (ProjectListResult, error) {
-	limit := options.Limit
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-	offset := options.Offset
-	if offset < 0 {
-		offset = 0
-	}
-	rows, err := s.db.QueryContext(ctx, `SELECT id, name, source_path, source_json, current_revision, spec_hash, created_at, updated_at, removed_at
-		FROM project ORDER BY updated_at DESC, created_at DESC, id ASC`)
-	if err != nil {
-		return ProjectListResult{}, fmt.Errorf("query projects: %w", err)
-	}
-	defer func() { _ = rows.Close() }()
-
-	query := strings.ToLower(strings.TrimSpace(options.Query))
-	matched := make([]ProjectRecord, 0)
-	for rows.Next() {
-		item, err := scanProject(rows.Scan)
-		if err != nil {
-			return ProjectListResult{}, err
-		}
-		if !options.IncludeRemoved && !item.RemovedAt.IsZero() {
-			continue
-		}
-		if query != "" && !projectMatchesQuery(item, query) {
-			continue
-		}
-		matched = append(matched, item)
-	}
-	if err := rows.Err(); err != nil {
-		return ProjectListResult{}, fmt.Errorf("iterate projects: %w", err)
-	}
-	total := len(matched)
-	end := offset + limit
-	if offset > total {
-		offset = total
-	}
-	if end > total {
-		end = total
-	}
-	return ProjectListResult{
-		Projects:   matched[offset:end],
-		TotalCount: total,
-		HasMore:    end < total,
-		NextOffset: end,
-	}, nil
+	return s.projectStore().ListProjects(ctx, options)
 }
 
 func (s *ConfigStore) GetProjectRevision(ctx context.Context, projectID string, revision int64) (ProjectRevisionRecord, error) {
-	row := s.db.QueryRowContext(ctx, `SELECT project_id, revision, spec_hash, spec_json, created_at
-		FROM project_revision WHERE project_id = ? AND revision = ?`, strings.TrimSpace(projectID), revision)
-	item, err := scanProjectRevision(row.Scan)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			return ProjectRevisionRecord{}, fmt.Errorf("project revision %s/%d not found: %w", strings.TrimSpace(projectID), revision, err)
-		}
-		return ProjectRevisionRecord{}, err
-	}
-	return item, nil
+	return s.projectStore().GetProjectRevision(ctx, projectID, revision)
 }
 
 func (s *ConfigStore) UpsertProjectAgent(ctx context.Context, agent ProjectAgentRecord) (ProjectAgentRecord, error) {
-	agent, err := normalizeProjectAgentRecord(agent)
-	if err != nil {
-		return ProjectAgentRecord{}, err
-	}
-	now := time.Now().UTC()
-	result, err := s.db.ExecContext(ctx, `UPDATE project_agent SET
-		managed_agent_id = ?, revision = ?, provider = ?, model = ?, image = ?, driver = ?, scheduler_enabled = ?, spec_json = ?, updated_at = ?
-		WHERE project_id = ? AND agent_name = ?`,
-		agent.ManagedAgentID, agent.Revision, agent.Provider, agent.Model, agent.Image, agent.Driver, boolToInt(agent.SchedulerEnabled), agent.SpecJSON, now.Unix(),
-		agent.ProjectID, agent.AgentName)
-	if err != nil {
-		return ProjectAgentRecord{}, fmt.Errorf("update project agent %s/%s: %w", agent.ProjectID, agent.AgentName, err)
-	}
-	if rows, _ := result.RowsAffected(); rows > 0 {
-		return s.GetProjectAgent(ctx, agent.ProjectID, agent.AgentName)
-	}
-	agent.CreatedAt = now
-	agent.UpdatedAt = now
-	if _, err := s.db.ExecContext(ctx, `INSERT INTO project_agent(
-		project_id, agent_name, managed_agent_id, revision, provider, model, image, driver, scheduler_enabled, spec_json, created_at, updated_at
-	) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-		agent.ProjectID, agent.AgentName, agent.ManagedAgentID, agent.Revision, agent.Provider, agent.Model, agent.Image, agent.Driver, boolToInt(agent.SchedulerEnabled), agent.SpecJSON,
-		agent.CreatedAt.Unix(), agent.UpdatedAt.Unix()); err != nil {
-		return ProjectAgentRecord{}, fmt.Errorf("insert project agent %s/%s: %w", agent.ProjectID, agent.AgentName, err)
-	}
-	return s.GetProjectAgent(ctx, agent.ProjectID, agent.AgentName)
+	return s.projectStore().UpsertProjectAgent(ctx, agent)
 }
 
 func (s *ConfigStore) GetProjectAgent(ctx context.Context, projectID, agentName string) (ProjectAgentRecord, error) {
-	row := s.db.QueryRowContext(ctx, `SELECT project_id, agent_name, managed_agent_id, revision, provider, model, image, driver, scheduler_enabled, spec_json, created_at, updated_at
-		FROM project_agent WHERE project_id = ? AND agent_name = ?`, strings.TrimSpace(projectID), strings.TrimSpace(agentName))
-	item, err := scanProjectAgent(row.Scan)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			return ProjectAgentRecord{}, fmt.Errorf("project agent %s/%s not found: %w", strings.TrimSpace(projectID), strings.TrimSpace(agentName), err)
-		}
-		return ProjectAgentRecord{}, err
-	}
-	return item, nil
+	return s.projectStore().GetProjectAgent(ctx, projectID, agentName)
 }
 
 func (s *ConfigStore) ListProjectAgents(ctx context.Context, projectID string) ([]ProjectAgentRecord, error) {
-	rows, err := s.db.QueryContext(ctx, `SELECT project_id, agent_name, managed_agent_id, revision, provider, model, image, driver, scheduler_enabled, spec_json, created_at, updated_at
-		FROM project_agent WHERE project_id = ? ORDER BY agent_name ASC`, strings.TrimSpace(projectID))
-	if err != nil {
-		return nil, fmt.Errorf("query project agents %s: %w", strings.TrimSpace(projectID), err)
-	}
-	defer func() { _ = rows.Close() }()
-	var items []ProjectAgentRecord
-	for rows.Next() {
-		item, err := scanProjectAgent(rows.Scan)
-		if err != nil {
-			return nil, err
-		}
-		items = append(items, item)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("iterate project agents %s: %w", strings.TrimSpace(projectID), err)
-	}
-	return items, nil
+	return s.projectStore().ListProjectAgents(ctx, projectID)
 }
 
 func (s *ConfigStore) UpsertProjectScheduler(ctx context.Context, scheduler ProjectSchedulerRecord) (ProjectSchedulerRecord, error) {
-	scheduler, err := normalizeProjectSchedulerRecord(scheduler)
-	if err != nil {
-		return ProjectSchedulerRecord{}, err
-	}
-	now := time.Now().UTC()
-	result, err := s.db.ExecContext(ctx, `UPDATE project_scheduler SET
-		agent_name = ?, managed_loader_id = ?, revision = ?, enabled = ?, trigger_count = ?, spec_json = ?, updated_at = ?
-		WHERE project_id = ? AND scheduler_id = ?`,
-		scheduler.AgentName, scheduler.ManagedLoaderID, scheduler.Revision, boolToInt(scheduler.Enabled), scheduler.TriggerCount, scheduler.SpecJSON, now.Unix(),
-		scheduler.ProjectID, scheduler.SchedulerID)
-	if err != nil {
-		return ProjectSchedulerRecord{}, fmt.Errorf("update project scheduler %s/%s: %w", scheduler.ProjectID, scheduler.SchedulerID, err)
-	}
-	if rows, _ := result.RowsAffected(); rows > 0 {
-		return s.GetProjectScheduler(ctx, scheduler.ProjectID, scheduler.SchedulerID)
-	}
-	scheduler.CreatedAt = now
-	scheduler.UpdatedAt = now
-	if _, err := s.db.ExecContext(ctx, `INSERT INTO project_scheduler(
-		project_id, scheduler_id, agent_name, managed_loader_id, revision, enabled, trigger_count, spec_json, created_at, updated_at
-	) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-		scheduler.ProjectID, scheduler.SchedulerID, scheduler.AgentName, scheduler.ManagedLoaderID, scheduler.Revision, boolToInt(scheduler.Enabled), scheduler.TriggerCount, scheduler.SpecJSON,
-		scheduler.CreatedAt.Unix(), scheduler.UpdatedAt.Unix()); err != nil {
-		return ProjectSchedulerRecord{}, fmt.Errorf("insert project scheduler %s/%s: %w", scheduler.ProjectID, scheduler.SchedulerID, err)
-	}
-	return s.GetProjectScheduler(ctx, scheduler.ProjectID, scheduler.SchedulerID)
+	return s.projectStore().UpsertProjectScheduler(ctx, scheduler)
 }
 
 func (s *ConfigStore) GetProjectScheduler(ctx context.Context, projectID, schedulerID string) (ProjectSchedulerRecord, error) {
-	row := s.db.QueryRowContext(ctx, `SELECT project_id, scheduler_id, agent_name, managed_loader_id, revision, enabled, trigger_count, spec_json, created_at, updated_at
-		FROM project_scheduler WHERE project_id = ? AND scheduler_id = ?`, strings.TrimSpace(projectID), strings.TrimSpace(schedulerID))
-	item, err := scanProjectScheduler(row.Scan)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			return ProjectSchedulerRecord{}, fmt.Errorf("project scheduler %s/%s not found: %w", strings.TrimSpace(projectID), strings.TrimSpace(schedulerID), err)
-		}
-		return ProjectSchedulerRecord{}, err
-	}
-	return item, nil
+	return s.projectStore().GetProjectScheduler(ctx, projectID, schedulerID)
 }
 
 func (s *ConfigStore) SetProjectSchedulerEnabled(ctx context.Context, projectID, schedulerID string, enabled bool) (ProjectSchedulerRecord, error) {
-	projectID = strings.TrimSpace(projectID)
-	schedulerID = strings.TrimSpace(schedulerID)
-	if projectID == "" || schedulerID == "" {
-		return ProjectSchedulerRecord{}, fmt.Errorf("project scheduler id is required")
-	}
-	result, err := s.db.ExecContext(ctx, `UPDATE project_scheduler SET enabled = ?, updated_at = ? WHERE project_id = ? AND scheduler_id = ?`,
-		boolToInt(enabled), time.Now().UTC().Unix(), projectID, schedulerID)
-	if err != nil {
-		return ProjectSchedulerRecord{}, fmt.Errorf("update project scheduler %s/%s enabled state: %w", projectID, schedulerID, err)
-	}
-	if rows, _ := result.RowsAffected(); rows == 0 {
-		return ProjectSchedulerRecord{}, fmt.Errorf("project scheduler %s/%s not found", projectID, schedulerID)
-	}
-	return s.GetProjectScheduler(ctx, projectID, schedulerID)
+	return s.projectStore().SetProjectSchedulerEnabled(ctx, projectID, schedulerID, enabled)
 }
 
 func (s *ConfigStore) ListProjectSchedulers(ctx context.Context, projectID string) ([]ProjectSchedulerRecord, error) {
-	rows, err := s.db.QueryContext(ctx, `SELECT project_id, scheduler_id, agent_name, managed_loader_id, revision, enabled, trigger_count, spec_json, created_at, updated_at
-		FROM project_scheduler WHERE project_id = ? ORDER BY agent_name ASC, scheduler_id ASC`, strings.TrimSpace(projectID))
-	if err != nil {
-		return nil, fmt.Errorf("query project schedulers %s: %w", strings.TrimSpace(projectID), err)
-	}
-	defer func() { _ = rows.Close() }()
-	var items []ProjectSchedulerRecord
-	for rows.Next() {
-		item, err := scanProjectScheduler(rows.Scan)
-		if err != nil {
-			return nil, err
-		}
-		items = append(items, item)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("iterate project schedulers %s: %w", strings.TrimSpace(projectID), err)
-	}
-	return items, nil
+	return s.projectStore().ListProjectSchedulers(ctx, projectID)
 }
 
 func (s *ConfigStore) CreateProjectRun(ctx context.Context, run ProjectRunRecord) (ProjectRunRecord, error) {
-	run, err := normalizeProjectRunRecord(run)
-	if err != nil {
-		return ProjectRunRecord{}, err
-	}
-	now := time.Now().UTC()
-	run.CreatedAt = now
-	run.UpdatedAt = now
-	if _, err := s.db.ExecContext(ctx, `INSERT INTO project_run(
-		run_id, project_id, project_name, project_revision, agent_name, managed_agent_id, source, scheduler_id, trigger_id, status,
-		session_id, exit_code, error, prompt, output, result_json, logs_path, artifacts_dir, cleanup_error, driver, image_ref,
-		started_at, completed_at, duration_ms, created_at, updated_at
-	) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-		run.RunID, run.ProjectID, run.ProjectName, run.ProjectRevision, run.AgentName, run.ManagedAgentID, run.Source, run.SchedulerID, run.TriggerID, run.Status,
-		run.SessionID, run.ExitCode, run.Error, run.Prompt, run.Output, run.ResultJSON, run.LogsPath, run.ArtifactsDir, run.CleanupError, run.Driver, run.ImageRef,
-		nonZeroTimeUnixMilli(run.StartedAt), nonZeroTimeUnixMilli(run.CompletedAt), run.DurationMs, run.CreatedAt.Unix(), run.UpdatedAt.Unix()); err != nil {
-		return ProjectRunRecord{}, fmt.Errorf("insert project run %s: %w", run.RunID, err)
-	}
-	return s.GetProjectRun(ctx, run.RunID)
+	return s.projectStore().CreateProjectRun(ctx, run)
 }
 
 func (s *ConfigStore) UpdateProjectRun(ctx context.Context, run ProjectRunRecord) (ProjectRunRecord, error) {
-	run, err := normalizeProjectRunRecord(run)
-	if err != nil {
-		return ProjectRunRecord{}, err
-	}
-	now := time.Now().UTC()
-	result, err := s.db.ExecContext(ctx, `UPDATE project_run SET
-		project_id = ?, project_name = ?, project_revision = ?, agent_name = ?, managed_agent_id = ?, source = ?, scheduler_id = ?, trigger_id = ?, status = ?,
-		session_id = ?, exit_code = ?, error = ?, prompt = ?, output = ?, result_json = ?, logs_path = ?, artifacts_dir = ?, cleanup_error = ?, driver = ?, image_ref = ?,
-		started_at = ?, completed_at = ?, duration_ms = ?, updated_at = ?
-		WHERE run_id = ?`,
-		run.ProjectID, run.ProjectName, run.ProjectRevision, run.AgentName, run.ManagedAgentID, run.Source, run.SchedulerID, run.TriggerID, run.Status,
-		run.SessionID, run.ExitCode, run.Error, run.Prompt, run.Output, run.ResultJSON, run.LogsPath, run.ArtifactsDir, run.CleanupError, run.Driver, run.ImageRef,
-		nonZeroTimeUnixMilli(run.StartedAt), nonZeroTimeUnixMilli(run.CompletedAt), run.DurationMs, now.Unix(), run.RunID)
-	if err != nil {
-		return ProjectRunRecord{}, fmt.Errorf("update project run %s: %w", run.RunID, err)
-	}
-	if rows, _ := result.RowsAffected(); rows == 0 {
-		return ProjectRunRecord{}, fmt.Errorf("project run %s not found", run.RunID)
-	}
-	return s.GetProjectRun(ctx, run.RunID)
+	return s.projectStore().UpdateProjectRun(ctx, run)
 }
 
 func (s *ConfigStore) GetProjectRun(ctx context.Context, runID string) (ProjectRunRecord, error) {
-	row := s.db.QueryRowContext(ctx, selectProjectRunSQL()+` WHERE run_id = ?`, strings.TrimSpace(runID))
-	item, err := scanProjectRun(row.Scan)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			return ProjectRunRecord{}, fmt.Errorf("project run %s not found: %w", strings.TrimSpace(runID), err)
-		}
-		return ProjectRunRecord{}, err
-	}
-	return item, nil
+	return s.projectStore().GetProjectRun(ctx, runID)
 }
 
 func (s *ConfigStore) ListProjectRuns(ctx context.Context, projectID string, limit int) ([]ProjectRunRecord, error) {
-	return s.ListProjectRunsByOptions(ctx, ProjectRunListOptions{ProjectID: projectID, Limit: limit})
+	return s.projectStore().ListProjectRuns(ctx, projectID, limit)
 }
 
 func (s *ConfigStore) ListProjectRunsByOptions(ctx context.Context, options ProjectRunListOptions) ([]ProjectRunRecord, error) {
-	limit := options.Limit
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-	offset := options.Offset
-	if offset < 0 {
-		offset = 0
-	}
-	where := make([]string, 0, 6)
-	args := make([]any, 0, 8)
-	if projectID := strings.TrimSpace(options.ProjectID); projectID != "" {
-		where = append(where, "project_id = ?")
-		args = append(args, projectID)
-	}
-	if agentName := strings.TrimSpace(options.AgentName); agentName != "" {
-		where = append(where, "agent_name = ?")
-		args = append(args, agentName)
-	}
-	if sessionID := strings.TrimSpace(options.SessionID); sessionID != "" {
-		where = append(where, "session_id = ?")
-		args = append(args, sessionID)
-	}
-	if schedulerID := strings.TrimSpace(options.SchedulerID); schedulerID != "" {
-		where = append(where, "scheduler_id = ?")
-		args = append(args, schedulerID)
-	}
-	if status := strings.TrimSpace(options.Status); status != "" {
-		where = append(where, "status = ?")
-		args = append(args, normalizeProjectRunStatus(status))
-	}
-	if source := strings.TrimSpace(options.Source); source != "" {
-		where = append(where, "source = ?")
-		args = append(args, normalizeProjectRunSource(source))
-	}
-	query := selectProjectRunSQL()
-	if len(where) > 0 {
-		query += ` WHERE ` + strings.Join(where, ` AND `)
-	}
-	query += ` ORDER BY created_at DESC, run_id DESC LIMIT ? OFFSET ?`
-	args = append(args, limit, offset)
-	rows, err := s.db.QueryContext(ctx, query, args...)
-	if err != nil {
-		return nil, fmt.Errorf("query project runs: %w", err)
-	}
-	defer func() { _ = rows.Close() }()
-	var items []ProjectRunRecord
-	for rows.Next() {
-		item, err := scanProjectRun(rows.Scan)
-		if err != nil {
-			return nil, err
-		}
-		items = append(items, item)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("iterate project runs: %w", err)
-	}
-	return items, nil
+	return s.projectStore().ListProjectRunsByOptions(ctx, options)
 }
 
 func (s *ConfigStore) getProject(ctx context.Context, projectID string, includeRemoved bool) (ProjectRecord, bool, error) {
-	projectID = strings.TrimSpace(projectID)
-	if projectID == "" {
-		return ProjectRecord{}, false, fmt.Errorf("project id is required")
-	}
-	where := "id = ? AND removed_at = 0"
 	if includeRemoved {
-		where = "id = ?"
+		return s.projectStore().GetProjectIncludingRemoved(ctx, projectID)
 	}
-	row := s.db.QueryRowContext(ctx, `SELECT id, name, source_path, source_json, current_revision, spec_hash, created_at, updated_at, removed_at
-		FROM project WHERE `+where, projectID)
-	item, err := scanProject(row.Scan)
+	project, err := s.GetProject(ctx, projectID)
 	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			return ProjectRecord{}, false, nil
-		}
 		return ProjectRecord{}, false, err
 	}
-	return item, true, nil
-}
-
-func normalizeProjectRecord(project ProjectRecord) (ProjectRecord, error) {
-	project.ID = strings.TrimSpace(project.ID)
-	project.Name = strings.TrimSpace(project.Name)
-	project.SourcePath = normalizeProjectSourcePath(project.SourcePath)
-	project.SourceJSON = strings.TrimSpace(project.SourceJSON)
-	project.SpecHash = strings.TrimSpace(project.SpecHash)
-	if project.ID == "" {
-		return ProjectRecord{}, fmt.Errorf("project id is required")
-	}
-	if project.Name == "" {
-		return ProjectRecord{}, fmt.Errorf("project name is required")
-	}
-	if project.SourceJSON == "" {
-		sourceJSON, err := encodeProjectSourceJSON(project.SourcePath)
-		if err != nil {
-			return ProjectRecord{}, err
-		}
-		project.SourceJSON = sourceJSON
-	}
-	if !json.Valid([]byte(project.SourceJSON)) {
-		return ProjectRecord{}, fmt.Errorf("project source_json must be valid JSON")
-	}
-	if project.CurrentRevision < 0 {
-		return ProjectRecord{}, fmt.Errorf("project current revision cannot be negative")
-	}
-	return project, nil
-}
-
-func normalizeProjectAgentRecord(agent ProjectAgentRecord) (ProjectAgentRecord, error) {
-	agent.ProjectID = strings.TrimSpace(agent.ProjectID)
-	agent.AgentName = strings.TrimSpace(agent.AgentName)
-	agent.ManagedAgentID = strings.TrimSpace(agent.ManagedAgentID)
-	agent.Provider = strings.TrimSpace(agent.Provider)
-	agent.Model = strings.TrimSpace(agent.Model)
-	agent.Image = strings.TrimSpace(agent.Image)
-	agent.Driver = strings.TrimSpace(agent.Driver)
-	agent.SpecJSON = strings.TrimSpace(agent.SpecJSON)
-	if agent.ProjectID == "" || agent.AgentName == "" {
-		return ProjectAgentRecord{}, fmt.Errorf("project id and agent name are required")
-	}
-	if agent.ManagedAgentID == "" {
-		managedAgentID, err := StableManagedAgentID(agent.ProjectID, agent.AgentName)
-		if err != nil {
-			return ProjectAgentRecord{}, err
-		}
-		agent.ManagedAgentID = managedAgentID
-	}
-	if agent.Revision < 0 {
-		return ProjectAgentRecord{}, fmt.Errorf("project agent revision cannot be negative")
-	}
-	if agent.SpecJSON == "" {
-		agent.SpecJSON = "{}"
-	}
-	if !json.Valid([]byte(agent.SpecJSON)) {
-		return ProjectAgentRecord{}, fmt.Errorf("project agent spec_json must be valid JSON")
-	}
-	return agent, nil
-}
-
-func normalizeProjectSchedulerRecord(scheduler ProjectSchedulerRecord) (ProjectSchedulerRecord, error) {
-	scheduler.ProjectID = strings.TrimSpace(scheduler.ProjectID)
-	scheduler.SchedulerID = strings.TrimSpace(scheduler.SchedulerID)
-	scheduler.AgentName = strings.TrimSpace(scheduler.AgentName)
-	scheduler.ManagedLoaderID = strings.TrimSpace(scheduler.ManagedLoaderID)
-	scheduler.SpecJSON = strings.TrimSpace(scheduler.SpecJSON)
-	if scheduler.ProjectID == "" || scheduler.AgentName == "" {
-		return ProjectSchedulerRecord{}, fmt.Errorf("project id and agent name are required")
-	}
-	if scheduler.SchedulerID == "" {
-		schedulerID, err := StableProjectSchedulerID(scheduler.ProjectID, scheduler.AgentName, "")
-		if err != nil {
-			return ProjectSchedulerRecord{}, err
-		}
-		scheduler.SchedulerID = schedulerID
-	}
-	if scheduler.ManagedLoaderID == "" {
-		loaderID, err := StableManagedLoaderID(scheduler.ProjectID, scheduler.AgentName, "")
-		if err != nil {
-			return ProjectSchedulerRecord{}, err
-		}
-		scheduler.ManagedLoaderID = loaderID
-	}
-	if scheduler.Revision < 0 {
-		return ProjectSchedulerRecord{}, fmt.Errorf("project scheduler revision cannot be negative")
-	}
-	if scheduler.TriggerCount < 0 {
-		return ProjectSchedulerRecord{}, fmt.Errorf("project scheduler trigger count cannot be negative")
-	}
-	if scheduler.SpecJSON == "" {
-		scheduler.SpecJSON = "{}"
-	}
-	if !json.Valid([]byte(scheduler.SpecJSON)) {
-		return ProjectSchedulerRecord{}, fmt.Errorf("project scheduler spec_json must be valid JSON")
-	}
-	return scheduler, nil
-}
-
-func normalizeProjectRunRecord(run ProjectRunRecord) (ProjectRunRecord, error) {
-	run.RunID = strings.TrimSpace(run.RunID)
-	run.ProjectID = strings.TrimSpace(run.ProjectID)
-	run.ProjectName = strings.TrimSpace(run.ProjectName)
-	run.AgentName = strings.TrimSpace(run.AgentName)
-	run.ManagedAgentID = strings.TrimSpace(run.ManagedAgentID)
-	run.Source = strings.TrimSpace(run.Source)
-	run.SchedulerID = strings.TrimSpace(run.SchedulerID)
-	run.TriggerID = strings.TrimSpace(run.TriggerID)
-	run.Status = normalizeProjectRunStatus(run.Status)
-	run.SessionID = strings.TrimSpace(run.SessionID)
-	run.ResultJSON = strings.TrimSpace(run.ResultJSON)
-	run.LogsPath = strings.TrimSpace(run.LogsPath)
-	run.ArtifactsDir = strings.TrimSpace(run.ArtifactsDir)
-	run.Driver = strings.TrimSpace(run.Driver)
-	run.ImageRef = strings.TrimSpace(run.ImageRef)
-	if run.RunID == "" || run.ProjectID == "" {
-		return ProjectRunRecord{}, fmt.Errorf("project run id and project id are required")
-	}
-	if run.ProjectRevision < 0 {
-		return ProjectRunRecord{}, fmt.Errorf("project run revision cannot be negative")
-	}
-	if run.ResultJSON == "" {
-		run.ResultJSON = "{}"
-	}
-	if !json.Valid([]byte(run.ResultJSON)) {
-		return ProjectRunRecord{}, fmt.Errorf("project run result_json must be valid JSON")
-	}
-	return run, nil
-}
-
-func normalizeProjectRunStatus(status string) string {
-	return domain.NormalizeProjectRunStatus(status)
-}
-
-func scanProject(scan func(dest ...any) error) (ProjectRecord, error) {
-	var item ProjectRecord
-	var createdAtRaw any
-	var updatedAtRaw any
-	var removedAtRaw any
-	if err := scan(&item.ID, &item.Name, &item.SourcePath, &item.SourceJSON, &item.CurrentRevision, &item.SpecHash, &createdAtRaw, &updatedAtRaw, &removedAtRaw); err != nil {
-		return ProjectRecord{}, fmt.Errorf("scan project: %w", err)
-	}
-	item.CreatedAt = parseStoredTime(createdAtRaw)
-	item.UpdatedAt = parseStoredTime(updatedAtRaw)
-	item.RemovedAt = parseStoredTime(removedAtRaw)
-	return item, nil
-}
-
-func scanProjectRevision(scan func(dest ...any) error) (ProjectRevisionRecord, error) {
-	var item ProjectRevisionRecord
-	var createdAtRaw any
-	if err := scan(&item.ProjectID, &item.Revision, &item.SpecHash, &item.SpecJSON, &createdAtRaw); err != nil {
-		return ProjectRevisionRecord{}, fmt.Errorf("scan project revision: %w", err)
-	}
-	item.CreatedAt = parseStoredTime(createdAtRaw)
-	return item, nil
-}
-
-func scanProjectAgent(scan func(dest ...any) error) (ProjectAgentRecord, error) {
-	var item ProjectAgentRecord
-	var schedulerEnabled int
-	var createdAtRaw any
-	var updatedAtRaw any
-	if err := scan(&item.ProjectID, &item.AgentName, &item.ManagedAgentID, &item.Revision, &item.Provider, &item.Model, &item.Image, &item.Driver, &schedulerEnabled, &item.SpecJSON, &createdAtRaw, &updatedAtRaw); err != nil {
-		return ProjectAgentRecord{}, fmt.Errorf("scan project agent: %w", err)
-	}
-	item.SchedulerEnabled = schedulerEnabled != 0
-	item.CreatedAt = parseStoredTime(createdAtRaw)
-	item.UpdatedAt = parseStoredTime(updatedAtRaw)
-	return item, nil
-}
-
-func scanProjectScheduler(scan func(dest ...any) error) (ProjectSchedulerRecord, error) {
-	var item ProjectSchedulerRecord
-	var enabled int
-	var createdAtRaw any
-	var updatedAtRaw any
-	if err := scan(&item.ProjectID, &item.SchedulerID, &item.AgentName, &item.ManagedLoaderID, &item.Revision, &enabled, &item.TriggerCount, &item.SpecJSON, &createdAtRaw, &updatedAtRaw); err != nil {
-		return ProjectSchedulerRecord{}, fmt.Errorf("scan project scheduler: %w", err)
-	}
-	item.Enabled = enabled != 0
-	item.CreatedAt = parseStoredTime(createdAtRaw)
-	item.UpdatedAt = parseStoredTime(updatedAtRaw)
-	return item, nil
-}
-
-func scanProjectRun(scan func(dest ...any) error) (ProjectRunRecord, error) {
-	var item ProjectRunRecord
-	var startedAtRaw any
-	var completedAtRaw any
-	var createdAtRaw any
-	var updatedAtRaw any
-	if err := scan(
-		&item.RunID, &item.ProjectID, &item.ProjectName, &item.ProjectRevision, &item.AgentName, &item.ManagedAgentID, &item.Source, &item.SchedulerID, &item.TriggerID, &item.Status,
-		&item.SessionID, &item.ExitCode, &item.Error, &item.Prompt, &item.Output, &item.ResultJSON, &item.LogsPath, &item.ArtifactsDir, &item.CleanupError, &item.Driver, &item.ImageRef,
-		&startedAtRaw, &completedAtRaw, &item.DurationMs, &createdAtRaw, &updatedAtRaw,
-	); err != nil {
-		return ProjectRunRecord{}, fmt.Errorf("scan project run: %w", err)
-	}
-	item.StartedAt = parseStoredUnixTimeAuto(asInt64Time(startedAtRaw))
-	item.CompletedAt = parseStoredUnixTimeAuto(asInt64Time(completedAtRaw))
-	item.CreatedAt = parseStoredTime(createdAtRaw)
-	item.UpdatedAt = parseStoredTime(updatedAtRaw)
-	return item, nil
-}
-
-func selectProjectRunSQL() string {
-	return `SELECT run_id, project_id, project_name, project_revision, agent_name, managed_agent_id, source, scheduler_id, trigger_id, status,
-		session_id, exit_code, error, prompt, output, result_json, logs_path, artifacts_dir, cleanup_error, driver, image_ref,
-		started_at, completed_at, duration_ms, created_at, updated_at FROM project_run`
-}
-
-func projectMatchesQuery(item ProjectRecord, query string) bool {
-	return strings.Contains(strings.ToLower(item.ID), query) ||
-		strings.Contains(strings.ToLower(item.Name), query) ||
-		strings.Contains(strings.ToLower(item.SourcePath), query)
-}
-
-func normalizeProjectSourcePath(sourcePath string) string {
-	return domain.NormalizeProjectSourcePath(sourcePath)
-}
-
-func encodeProjectSourceJSON(sourcePath string) (string, error) {
-	data, err := json.Marshal(struct {
-		ComposePath string `json:"compose_path,omitempty"`
-	}{ComposePath: normalizeProjectSourcePath(sourcePath)})
-	if err != nil {
-		return "", fmt.Errorf("marshal project source: %w", err)
-	}
-	return string(data), nil
-}
-
-func marshalCanonicalProjectJSON(value any) ([]byte, error) {
-	data, err := json.Marshal(value)
-	if err != nil {
-		return nil, err
-	}
-	return data, nil
+	return project, true, nil
 }
 
 func stableReadableID(prefix, readable, seed string) string {
@@ -936,37 +156,14 @@ func isProjectStableIdentifier(value string) bool {
 	return domain.IsProjectStableIdentifier(value)
 }
 
-func asInt64Time(value any) int64 {
-	switch typed := value.(type) {
-	case nil:
-		return 0
-	case int64:
-		return typed
-	case int:
-		return int64(typed)
-	case float64:
-		return int64(typed)
-	case []byte:
-		return asInt64Time(string(typed))
-	case string:
-		parsed, _ := parseInt64String(typed)
-		return parsed
-	default:
-		return 0
-	}
+func normalizeProjectRunStatus(status string) string {
+	return projectdomain.NormalizeProjectRunStatus(status)
 }
 
-func parseInt64String(value string) (int64, bool) {
-	value = strings.TrimSpace(value)
-	if value == "" {
-		return 0, false
-	}
-	var parsed int64
-	for _, r := range value {
-		if r < '0' || r > '9' {
-			return 0, false
-		}
-		parsed = parsed*10 + int64(r-'0')
-	}
-	return parsed, true
+func scanProjectRun(scan func(dest ...any) error) (ProjectRunRecord, error) {
+	return projectdomain.ScanProjectRun(scan)
+}
+
+func selectProjectRunSQL() string {
+	return projectdomain.SelectProjectRunSQL()
 }
